fix response on request get Campus

+++ src/main/java/com/yaetotui/backend_web_site/domain/entity/Coordinates.java
@@ -16,8 +16,8 @@ import lombok.Setter;
 @AllArgsConstructor
 @Table(name = "coordinates")
 public class Coordinates extends BaseDomainEntity {
-    float x;
-    float y;
+    Integer x;
+    Integer y;
 
     @ManyToOne()
     @JoinColumn(name = "cabinet_id")
